extract a messages package
