+++ deployables/extension/src/inject/bridge/useProviderBridge.spec.tsx
@@ -1,4 +1,3 @@
-import { InjectedProviderMessageTyp } from '@/messages'
 import {
   chromeMock,
   createMockTab,
@@ -10,6 +9,7 @@ import {
 import type { Eip1193Provider } from '@/types'
 import { cleanup, waitFor } from '@testing-library/react'
 import { ZERO_ADDRESS } from '@zodiac/chains'
+import { InjectedProviderMessageTyp } from '@zodiac/messages'
 import { toQuantity } from 'ethers'
 import type { PropsWithChildren } from 'react'
 import type { ChainId } from 'ser-kit'
