+++ deployables/extension/tsconfig.json
@@ -13,7 +13,6 @@
       "@/inject-bridge": ["./src/inject/bridge/index.ts"],
       "@/e2e-utils": ["./e2e/utils/index.ts"],
       "@/test-utils": ["./test-utils/index.ts"],
-      "@/messages": ["./src/messages.ts"],
       "@/transaction-translation": [
         "./src/panel/transactionTranslations/index.ts"
       ],
