+++ packages/messages/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "@zodiac/messages",
+  "private": true,
+  "type": "module",
+  "license": "UNLICENSED",
+  "sideEffects": false,
+  "packageManager": "pnpm@9.15.4",
+  "scripts": {
+    "lint": "eslint . --max-warnings=0",
+    "check-types": "tsc"
+  },
+  "exports": {
+    ".": {
+      "types": "./src/index.ts",
+      "import": "./src/index.ts"
+    }
+  },
+  "dependencies": {
+    "@zodiac/chains": "workspace:*",
+    "@zodiac/schema": "workspace:*"
+  },
+  "devDependencies": {
+    "@zodiac/eslint-config": "workspace:*",
+    "@zodiac/typescript-config": "workspace:*",
+    "eslint": "^9.7.0"
+  }
+}
