+++ packages/messages/src/JsonRpcRequest.ts
@@ -1,0 +1,4 @@
+export interface JsonRpcRequest {
+  method: string
+  params?: readonly unknown[] | object
+}
