+++ packages/messages/src/companionApp/index.ts
@@ -1,0 +1,1 @@
+export { CompanionAppMessageType, type CompanionAppMessage } from './editRoute'
