+++ packages/messages/src/extension/index.ts
@@ -1,0 +1,15 @@
+export { PilotMessageType, type Message } from './connect'
+export {
+  ConnectedWalletMessageType,
+  type ConnectedWalletMessage,
+} from './connectedWallet'
+export {
+  InjectedProviderMessageTyp,
+  type InjectedProviderMessage,
+  type InjectedProviderResponse,
+} from './injectedProvider'
+export { RpcMessageType, type RpcMessage } from './rpc'
+export {
+  PilotSimulationMessageType,
+  type SimulationMessage,
+} from './simulation'
