+++ packages/messages/src/extension/rpc.ts
@@ -1,0 +1,10 @@
+export enum RpcMessageType {
+  PROBE_CHAIN_ID = 'PROBE_CHAIN_ID',
+}
+
+interface ProbeChainId {
+  type: RpcMessageType.PROBE_CHAIN_ID
+  url: string
+}
+
+export type RpcMessage = ProbeChainId
