+++ packages/messages/src/extension/simulation.ts
@@ -1,0 +1,27 @@
+import { type ChainId } from '@zodiac/chains'
+
+export enum PilotSimulationMessageType {
+  SIMULATE_START = 'SIMULATE_START',
+  SIMULATE_UPDATE = 'SIMULATE_UPDATE',
+  SIMULATE_STOP = 'SIMULATE_STOP',
+}
+
+interface SimulateStart {
+  type: PilotSimulationMessageType.SIMULATE_START
+  windowId: number
+  chainId: ChainId
+  rpcUrl?: string
+}
+
+type SimulateUpdate = {
+  type: PilotSimulationMessageType.SIMULATE_UPDATE
+  windowId: number
+  rpcUrl: string
+}
+
+interface SimulateStop {
+  type: PilotSimulationMessageType.SIMULATE_STOP
+  windowId: number
+}
+
+export type SimulationMessage = SimulateStart | SimulateStop | SimulateUpdate
