+++ packages/messages/src/index.ts
@@ -1,0 +1,3 @@
+export * from './companionApp'
+export * from './extension'
+export { type JsonRpcRequest } from './JsonRpcRequest'
